// src/modules/user/enum/exceptions.enum.ts
export enum UserExceptions {
  EMAIL_ALREADY_EXISTS = "Email already registered",
  NOT_FOUND = "User not found",
  INVALID_EMAIL_OR_PASSWORD = "Invalid email or password",
  OUTDATED_AUTH_TOKEN = "Outdated auth token",
  ACCESS_DENIED_BY_REQUESTER_ROLE = "Access denied",
  BLOCKED = "Your account is temporarily blocked.",
}